test(IDX): isolate upgrade_existing_sns_test
The `//rs/nervous_system/integration_tests:integration_tests_test_tests/sns_ledger_upgrade` is our 3rd most flaky test.
To address it we isolate the biggest of its two subtests to its own dedicated bazel target and add logs so we'll know where the test is timing out.
