+++ rs/nervous_system/integration_tests/BUILD.bazel
@@ -173,6 +173,7 @@ rust_test_suite_with_extra_srcs(
             "tests/sns_upgrade_test_utils.rs",
             "tests/upgrade_everything_test.rs",
             "tests/advance_target_version_upgrades_all_canisters_test.rs",
+            "tests/upgrade_existing_sns_test.rs",
         ],
     ),
     aliases = ALIASES,
@@ -209,6 +210,22 @@ rust_test(
     name = "advance_target_version_upgrades_all_canisters_test",
     srcs = [
         "tests/advance_target_version_upgrades_all_canisters_test.rs",
+    ],
+    aliases = ALIASES,
+    data = DEV_DATA,
+    env = DEV_ENV,
+    flaky = True,
+    proc_macro_deps = MACRO_DEPENDENCIES + MACRO_DEV_DEPENDENCIES,
+    tags = [
+        "cpu:4",
+    ],
+    deps = [":nervous_system_integration_tests"] + DEPENDENCIES_WITH_TEST_FEATURES + DEV_DEPENDENCIES,
+)
+
+rust_test(
+    name = "upgrade_existing_sns_test",
+    srcs = [
+        "tests/upgrade_existing_sns_test.rs",
     ],
     aliases = ALIASES,
     data = DEV_DATA,
